minor changes trying to build with Eclipse CDT
git-svn-id: svn://svn.code.sf.net/p/suneido/svn/trunk@1139 8ef86ec9-3220-0410-98d7-ddd3c8784f88

+++ type.h
@@ -76,7 +76,7 @@ public:
 		{ return (T) n; }
 	};
 
-template<> class TypeInt<int64> : public Type
+template<> class TypeInt<long long> : public Type
 	{
 public:
 	int size()
